[testing] Split up model conversion test targets in CI (#1219)
