+++ tests/functional_tests/models/deepseek/test_deepseek_conversion.py
@@ -121,7 +121,9 @@ class TestDeepSeekConversion:
             str(ep),
         ]
 
-        result = subprocess.run(cmd, capture_output=True, text=True, cwd=Path(__file__).parent.parent.parent.parent)
+        result = subprocess.run(
+            cmd, capture_output=True, text=True, cwd=Path(__file__).parent.parent.parent.parent.parent
+        )
 
         if result.returncode != 0:
             print(f"STDOUT: {result.stdout}")
