+++ tests/functional_tests/models/deepseek/test_deepseek_provider_mapping.py
