+++ tests/functional_tests/models/gemma/test_gemma3_conversion.py
@@ -18,69 +18,49 @@ from pathlib import Path
 
 import pytest
 import torch
-from transformers import Qwen2_5_VLConfig, Qwen2_5_VLForConditionalGeneration, Qwen2Tokenizer
+from transformers import Gemma3ForCausalLM, Gemma3TextConfig, GemmaTokenizer
 
 
-HF_QWEN25_VL_TOY_MODEL_CONFIG = {
-    "architectures": ["Qwen2_5_VLForConditionalGeneration"],
+HF_GEMMA3_TOY_MODEL_CONFIG = {
+    "architectures": ["Gemma3ForCausalLM"],
+    "attention_bias": False,
     "attention_dropout": 0.0,
-    "bos_token_id": 151643,
-    "eos_token_id": 151645,
-    "vision_start_token_id": 151652,
-    "vision_end_token_id": 151653,
-    "vision_token_id": 151654,
-    "image_token_id": 151655,
-    "video_token_id": 151656,
-    "hidden_act": "silu",
-    "hidden_size": 3584,
+    "bos_token_id": 2,
+    "eos_token_id": 1,
+    "head_dim": 256,
+    "hidden_act": "gelu_pytorch_tanh",
+    "hidden_activation": "gelu_pytorch_tanh",
+    "hidden_size": 1152,  # Smaller than real 1B for faster testing
     "initializer_range": 0.02,
-    "intermediate_size": 18944,
-    "max_position_embeddings": 128000,
-    "max_window_layers": 2,
-    "model_type": "qwen2_5_vl",
-    "num_attention_heads": 28,
-    "num_hidden_layers": 2,
-    "num_key_value_heads": 4,
+    "intermediate_size": 2304,  # Reduced for TP compatibility testing
+    "max_position_embeddings": 8192,
+    "model_type": "gemma3_text",
+    "num_attention_heads": 4,
+    "num_hidden_layers": 2,  # Much smaller for testing
+    "num_key_value_heads": 2,  # Changed from 1 to 2 to be divisible by TP=2
+    "pad_token_id": 0,
+    "query_pre_attn_scalar": 256,
     "rms_norm_eps": 1e-06,
+    "rope_local_base_freq": 10000.0,
     "rope_theta": 1000000.0,
-    "sliding_window": 32768,
-    "tie_word_embeddings": False,
+    "sliding_window": 512,
     "torch_dtype": "bfloat16",
-    "transformers_version": "4.41.2",
+    "transformers_version": "4.46.0",
     "use_cache": True,
-    "use_sliding_window": False,
-    "vision_config": {
-        "depth": 2,
-        "hidden_act": "silu",
-        "hidden_size": 1280,
-        "intermediate_size": 3420,
-        "num_heads": 16,
-        "in_chans": 3,
-        "out_hidden_size": 3584,
-        "patch_size": 14,
-        "spatial_merge_size": 2,
-        "spatial_patch_size": 14,
-        "window_size": 112,
-        "fullatt_block_indexes": [
-            0,
-        ],
-        "tokens_per_second": 2,
-        "temporal_patch_size": 2,
-    },
-    "rope_scaling": {"type": "mrope", "mrope_section": [16, 24, 24]},
-    "vocab_size": 152064,
+    "vocab_size": 262144,
+    "rope_scaling": None,
 }
 
 
-class TestQwen25VLConversion:
+class TestGemma3Conversion:
     """
-    Test Qwen25 VL model conversion from local HuggingFace model with different parallelism configurations.
+    Test Gemma3 model conversion from local HuggingFace model with different parallelism configurations.
     """
 
     @pytest.fixture(scope="class")
-    def qwen25_vl_toy_model_path(self, tmp_path_factory):
+    def gemma3_toy_model_path(self, tmp_path_factory):
         """
-        Create and save a HuggingFace Qwen25 VL toy model from config to a temporary directory.
+        Create and save a HuggingFace Gemma3 toy model from config to a temporary directory.
 
         Args:
             tmp_path_factory: Pytest temporary path factory for class-scoped fixtures
@@ -89,15 +69,15 @@ class TestQwen25VLConversion:
             str: Path to the saved HuggingFace model directory
         """
         # Create a temporary directory for this test class
-        temp_dir = tmp_path_factory.mktemp("qwen25_vl_toy_model")
-        model_dir = temp_dir / "qwen25_vl_toy"
+        temp_dir = tmp_path_factory.mktemp("gemma3_toy_model")
+        model_dir = temp_dir / "gemma3_toy"
 
-        # Create Qwen25 VL config from the toy model config
-        config = Qwen2_5_VLConfig(**HF_QWEN25_VL_TOY_MODEL_CONFIG)
+        # Create Gemma3 config from the toy model config
+        config = Gemma3TextConfig(**HF_GEMMA3_TOY_MODEL_CONFIG)
         config.torch_dtype = torch.bfloat16  # Explicitly set the torch_dtype in config
 
         # Create model with random weights and convert to bfloat16
-        model = Qwen2_5_VLForConditionalGeneration(config)
+        model = Gemma3ForCausalLM(config)
         model = model.bfloat16()  # Use .bfloat16() method instead of .to()
 
         # Debug: Check model dtype before saving
@@ -105,47 +85,40 @@ class TestQwen25VLConversion:
             print(f"Before save - {name}: {param.dtype}")
             break  # Just check the first parameter
 
-        # Download and save tokenizer from a reference Qwen25 VL model
-        # We use the smallest available Qwen25 VL model for tokenizer artifacts
-        try:
-            tokenizer = Qwen2Tokenizer.from_pretrained("Qwen/Qwen2.5-VL-3B-Instruct")
-            tokenizer.save_pretrained(model_dir)
-        except Exception as e:
-            print(f"Warning: Could not download tokenizer, creating minimal tokenizer files: {e}")
-            # Create minimal tokenizer files if download fails
-            # This is a fallback for offline environments
-            tokenizer_config = {
-                "tokenizer_class": "Qwen2Tokenizer",
-                "vocab_size": 151936,
-                "bos_token": "<|endoftext|>",
-                "eos_token": "<|endoftext|>",
-                "pad_token": "<|endoftext|>",
-                "unk_token": "<|endoftext|>",
-            }
-
-            with open(model_dir / "tokenizer_config.json", "w") as f:
-                json.dump(tokenizer_config, f, indent=2)
+        # Download and save tokenizer from a reference Gemma model
+        # We use the smallest available Gemma model for tokenizer artifacts
+        # First try to load from pre-mounted test data, then fall back to HuggingFace download
+        pre_downloaded_path = "/home/TestData/megatron_bridge/tokenizers/google/gemma-2b"
+        # Try loading from pre-downloaded location first
+        if Path(pre_downloaded_path).exists():
+            print(f"Loading tokenizer from pre-downloaded path: {pre_downloaded_path}")
+            tokenizer = GemmaTokenizer.from_pretrained(pre_downloaded_path)
+        else:
+            # Fall back to downloading from HuggingFace
+            print("Pre-downloaded tokenizer not found, attempting to download from HuggingFace")
+            tokenizer = GemmaTokenizer.from_pretrained("google/gemma-2b")
+        tokenizer.save_pretrained(model_dir)
 
         # Save model and config to directory
         model.save_pretrained(model_dir, safe_serialization=True)
 
         # Also save config.json explicitly to ensure compatibility with correct torch_dtype
-        config_to_save = HF_QWEN25_VL_TOY_MODEL_CONFIG.copy()
+        config_to_save = HF_GEMMA3_TOY_MODEL_CONFIG.copy()
         config_path = model_dir / "config.json"
         with open(config_path, "w") as f:
             json.dump(config_to_save, f, indent=2)
 
         return str(model_dir)
 
-    def test_toy_model_creation(self, qwen25_vl_toy_model_path):
+    def test_toy_model_creation(self, gemma3_toy_model_path):
         """
         Test that the toy model is created correctly and can be loaded.
 
         Args:
-            qwen25_vl_toy_model_path: Path to the toy Qwen25 VL model (from fixture)
+            gemma3_toy_model_path: Path to the toy Gemma3 model (from fixture)
         """
         # Verify the model directory exists
-        model_path = Path(qwen25_vl_toy_model_path)
+        model_path = Path(gemma3_toy_model_path)
         assert model_path.exists(), f"Model directory not found at {model_path}"
 
         # Check essential files exist
@@ -166,36 +139,41 @@ class TestQwen25VLConversion:
         with open(config_file) as f:
             config_data = json.load(f)
 
-        assert config_data["model_type"] == "qwen2_5_vl"
-        assert config_data["hidden_size"] == 3584
+        assert config_data["model_type"] == "gemma3_text"
+        assert config_data["hidden_size"] == 1152
+        assert config_data["intermediate_size"] == 2304
         assert config_data["num_hidden_layers"] == 2
-        assert config_data["num_attention_heads"] == 28
-        assert config_data["vocab_size"] == 152064
-        assert "vision_config" in config_data
+        assert config_data["num_attention_heads"] == 4
+        assert config_data["num_key_value_heads"] == 2
+        assert config_data["vocab_size"] == 262144
+        assert config_data["head_dim"] == 256
+        # Check Gemma3-specific parameters
+        assert config_data["query_pre_attn_scalar"] == 256
+        assert config_data["sliding_window"] == 512
+        assert config_data["rope_local_base_freq"] == 10000.0
+        assert config_data["rope_theta"] == 1000000.0
 
         # Try loading the model to verify it's valid
         try:
-            model = Qwen2_5_VLForConditionalGeneration.from_pretrained(
-                qwen25_vl_toy_model_path,
+            model = Gemma3ForCausalLM.from_pretrained(
+                gemma3_toy_model_path,
                 torch_dtype=torch.bfloat16,
                 low_cpu_mem_usage=False,  # Ensure full loading
             )
 
             # Try loading the tokenizer as well
             try:
-                tokenizer = Qwen2Tokenizer.from_pretrained(qwen25_vl_toy_model_path)
+                tokenizer = GemmaTokenizer.from_pretrained(gemma3_toy_model_path)
                 print(f"Tokenizer loaded successfully with vocab_size: {tokenizer.vocab_size}")
             except Exception as e:
                 print(f"Warning: Could not load tokenizer (this might be OK for conversion testing): {e}")
 
-            # Verify model structure for VL model
+            # Verify model structure
             assert hasattr(model, "model")
-            assert hasattr(model.model, "language_model")
-            assert hasattr(model.model.language_model, "layers")
-            assert len(model.model.language_model.layers) == 2  # num_hidden_layers
-            assert hasattr(model.model, "visual")  # VL model should have visual component
+            assert hasattr(model.model, "layers")
+            assert len(model.model.layers) == 2  # num_hidden_layers
 
-            print(f"SUCCESS: Toy model created and validated at {qwen25_vl_toy_model_path}")
+            print(f"SUCCESS: Toy model created and validated at {gemma3_toy_model_path}")
             print("Model weights are correctly in bfloat16 format")
 
         except Exception as e:
@@ -209,23 +187,21 @@ class TestQwen25VLConversion:
             (1, 2, "PP"),
         ],
     )
-    def test_qwen25_vl_conversion_parallelism(self, qwen25_vl_toy_model_path, tmp_path, tp, pp, test_name):
+    def test_gemma3_conversion_parallelism(self, gemma3_toy_model_path, tmp_path, tp, pp, test_name):
         """
-        Test Qwen25 VL model conversion with different parallelism configurations.
+        Test Gemma3 model conversion with different parallelism configurations.
 
         Args:
-            qwen25_vl_toy_model_path: Path to the toy Qwen25 VL model (from fixture)
+            gemma3_toy_model_path: Path to the toy Gemma3 model (from fixture)
             tmp_path: Pytest temporary path fixture
             tp: Tensor parallelism size
             pp: Pipeline parallelism size
             test_name: Name of the test for identification
         """
-
         # Create temporary output directory for conversion results
-        test_output_dir = tmp_path / f"qwen25_vl_{test_name}"
+        test_output_dir = tmp_path / f"gemma3_{test_name}"
         test_output_dir.mkdir(exist_ok=True)
 
-        # Run hf_megatron_roundtrip_multi_gpu.py with specified parallelism configuration on our toy model
         cmd = [
             "python",
             "-m",
@@ -240,7 +216,7 @@ class TestQwen25VLConversion:
             "--parallel-mode",
             "examples/conversion/hf_megatron_roundtrip_multi_gpu.py",
             "--hf-model-id",
-            qwen25_vl_toy_model_path,  # Use our local toy model instead of downloading
+            gemma3_toy_model_path,
             "--output-dir",
             str(test_output_dir),
             "--tp",
@@ -251,19 +227,18 @@ class TestQwen25VLConversion:
 
         try:
             result = subprocess.run(
-                cmd, capture_output=True, text=True, cwd=Path(__file__).parent.parent.parent.parent
+                cmd, capture_output=True, text=True, cwd=Path(__file__).parent.parent.parent.parent.parent
             )
-            print(cmd)
 
             # Check that the conversion completed successfully
             if result.returncode != 0:
                 print(f"STDOUT: {result.stdout}")
                 print(f"STDERR: {result.stderr}")
-                assert False, f"Qwen25 VL {test_name} conversion failed with return code {result.returncode}"
+                assert False, f"Gemma3 {test_name} conversion failed with return code {result.returncode}"
 
             # Verify that the converted model was saved
             # The output directory should be named after the last part of the model path
-            model_name = Path(qwen25_vl_toy_model_path).name  # "qwen25_vl_toy"
+            model_name = Path(gemma3_toy_model_path).name  # "gemma3_toy"
             converted_model_dir = test_output_dir / model_name
             assert converted_model_dir.exists(), f"Converted model directory not found at {converted_model_dir}"
 
@@ -278,18 +253,25 @@ class TestQwen25VLConversion:
                 f"Model weights file not found in converted model at {converted_model_dir}"
             )
 
-            # Verify the config contains Qwen25 VL-specific parameters
+            # Verify the config contains Gemma3-specific parameters
             with open(config_file) as f:
                 saved_config = json.load(f)
 
-            assert saved_config["model_type"] == "qwen2_5_vl", "Model type should be qwen2_5_vl"
-            assert saved_config["hidden_size"] == 3584, "Hidden size should match toy config"
-            assert saved_config["num_attention_heads"] == 28, "Number of attention heads should match toy config"
-            assert "vision_config" in saved_config, "VL model should have vision_config"
+            assert saved_config["model_type"] == "gemma3_text", "Model type should be gemma3_text"
+            assert saved_config["hidden_size"] == 1152, "Hidden size should match toy config"
+            assert saved_config["intermediate_size"] == 2304, "Intermediate size should match toy config"
+            assert saved_config["num_attention_heads"] == 4, "Number of attention heads should match toy config"
+            assert saved_config["num_key_value_heads"] == 2, "Number of key-value heads should match toy config"
+            assert saved_config["head_dim"] == 256, "Head dimension should match toy config"
+            # Verify Gemma3-specific parameters
+            assert saved_config["query_pre_attn_scalar"] == 256, "Query pre-attention scalar should match"
+            assert saved_config["sliding_window"] == 512, "Sliding window should match"
+            assert saved_config["rope_local_base_freq"] == 10000.0, "Rope local base frequency should match"
+            assert saved_config["rope_theta"] == 1000000.0, "Rope theta should match"
 
-            print(f"SUCCESS: Qwen25 VL {test_name} conversion test completed successfully")
+            print(f"SUCCESS: Gemma3 {test_name} conversion test completed successfully")
             print(f"Converted model saved at: {converted_model_dir}")
 
         except Exception as e:
-            print(f"Error during Qwen25 VL {test_name} conversion test: {e}")
+            print(f"Error during Gemma3 {test_name} conversion test: {e}")
             raise
