+++ tests/functional_tests/models/gemma_vl/test_gemma3_vl_conversion.py
@@ -208,8 +208,8 @@ class TestGemma3VLConversion:
             "-m",
             "coverage",
             "run",
-            "--data-file=/workspace/.coverage",
-            "--source=/workspace/",
+            "--data-file=/opt/Megatron-Bridge/.coverage",
+            "--source=/opt/Megatron-Bridge/",
             "--parallel-mode",
             "examples/conversion/hf_megatron_roundtrip_multi_gpu.py",
             "--hf-model-id",
@@ -222,7 +222,9 @@ class TestGemma3VLConversion:
             str(pp),
         ]
 
-        result = subprocess.run(cmd, capture_output=True, text=True, cwd=Path(__file__).parent.parent.parent.parent)
+        result = subprocess.run(
+            cmd, capture_output=True, text=True, cwd=Path(__file__).parent.parent.parent.parent.parent
+        )
         print(cmd)
 
         # Check that the conversion completed successfully
