+++ tests/functional_tests/models/glm/test_glm45_conversion.py
@@ -18,56 +18,56 @@ from pathlib import Path
 
 import pytest
 import torch
-from transformers import AutoTokenizer, Qwen3MoeConfig, Qwen3MoeForCausalLM
+from transformers import AutoConfig, AutoTokenizer
 
 
-HF_QWEN3_MOE_TOY_MODEL_CONFIG = {
-    "architectures": ["Qwen3MoeForCausalLM"],
-    "attention_bias": False,
+HF_GLM45_TOY_MODEL_CONFIG = {
+    "architectures": ["Glm4MoeForCausalLM"],
+    "attention_bias": True,
     "attention_dropout": 0.0,
-    "bos_token_id": 151643,
-    "decoder_sparse_step": 1,
-    "eos_token_id": 151645,
+    "bos_token_id": 151329,
+    "eos_token_id": 151336,
+    "first_k_dense_replace": 1,
     "head_dim": 128,
     "hidden_act": "silu",
-    "hidden_size": 2048,
+    "hidden_size": 1024,
     "initializer_range": 0.02,
-    "intermediate_size": 6144,
-    "max_position_embeddings": 262144,
-    "max_window_layers": 48,
-    "mlp_only_layers": [],
-    "model_type": "qwen3_moe",
-    "moe_intermediate_size": 768,
+    "intermediate_size": 2048,
+    "max_position_embeddings": 8192,
+    "model_type": "glm",
+    "moe_intermediate_size": 512,
+    "n_routed_experts": 8,
+    "n_shared_experts": 1,
+    "n_group": 1,
+    "topk_group": 1,
     "norm_topk_prob": True,
-    "num_attention_heads": 32,
-    "num_experts": 4,
+    "num_attention_heads": 8,
     "num_experts_per_tok": 4,
     "num_hidden_layers": 2,
-    "num_key_value_heads": 4,
-    "output_router_logits": False,
+    "num_key_value_heads": 2,
+    "partial_rotary_factor": 0.5,
     "rms_norm_eps": 1e-06,
-    "rope_scaling": None,
-    "rope_theta": 10000000,
-    "router_aux_loss_coef": 0.001,
-    "sliding_window": None,
+    "rope_theta": 1000000.0,
+    "routed_scaling_factor": 2.5,
+    "num_nextn_predict_layers": 0,  # Huggingface initialization does not handle MTP correctly
     "tie_word_embeddings": False,
     "torch_dtype": "bfloat16",
-    "transformers_version": "4.51.0",
+    "transformers_version": "4.54.0",
     "use_cache": True,
-    "use_sliding_window": False,
-    "vocab_size": 151936,
+    "use_qk_norm": True,
+    "vocab_size": 151552,
 }
 
 
-class TestQwen3MoEConversion:
+class TestGLM45Conversion:
     """
-    Test Qwen3 MoE model conversion from local HuggingFace model with different parallelism configurations.
+    Test GLM 4.5 MoE model conversion from local HuggingFace model with different parallelism configurations.
     """
 
     @pytest.fixture(scope="class")
-    def qwen3_moe_toy_model_path(self, tmp_path_factory):
+    def glm45_toy_model_path(self, tmp_path_factory):
         """
-        Create and save a HuggingFace Qwen3 MoE toy model from config to a temporary directory.
+        Create and save a HuggingFace GLM 4.5 MoE toy model from config to a temporary directory.
 
         Args:
             tmp_path_factory: Pytest temporary path factory for class-scoped fixtures
@@ -76,15 +76,23 @@ class TestQwen3MoEConversion:
             str: Path to the saved HuggingFace model directory
         """
         # Create a temporary directory for this test class
-        temp_dir = tmp_path_factory.mktemp("qwen3_moe_toy_model")
-        model_dir = temp_dir / "qwen3_moe_toy"
+        temp_dir = tmp_path_factory.mktemp("glm45_toy_model")
+        model_dir = temp_dir / "glm45_toy"
 
-        # Create Qwen3 MoE config from the toy model config
-        config = Qwen3MoeConfig(**HF_QWEN3_MOE_TOY_MODEL_CONFIG)
+        # Create GLM 4.5 config from the toy model config using AutoConfig
+        config = AutoConfig.from_pretrained("zai-org/GLM-4.5")
+
+        # Override with toy model config
+        for key, value in HF_GLM45_TOY_MODEL_CONFIG.items():
+            setattr(config, key, value)
+
         config.torch_dtype = torch.bfloat16  # Explicitly set the torch_dtype in config
 
         # Create model with random weights and convert to bfloat16
-        model = Qwen3MoeForCausalLM(config)
+        from transformers import Glm4MoeForCausalLM
+
+        model = Glm4MoeForCausalLM(config)
+
         model = model.bfloat16()  # Use .bfloat16() method instead of .to()
 
         # Debug: Check model dtype before saving
@@ -92,30 +100,30 @@ class TestQwen3MoEConversion:
             print(f"Before save - {name}: {param.dtype}")
             break  # Just check the first parameter
 
-        # Download and save tokenizer from a reference Qwen model
-        tokenizer = AutoTokenizer.from_pretrained("Qwen/Qwen3-0.6B")
+        # Download and save tokenizer from a reference GLM model
+        tokenizer = AutoTokenizer.from_pretrained("zai-org/GLM-4.5")
         tokenizer.save_pretrained(model_dir)
 
         # Save model and config to directory
         model.save_pretrained(model_dir, safe_serialization=True)
 
         # Also save config.json explicitly to ensure compatibility with correct torch_dtype
-        config_to_save = HF_QWEN3_MOE_TOY_MODEL_CONFIG.copy()
+        config_to_save = HF_GLM45_TOY_MODEL_CONFIG.copy()
         config_path = model_dir / "config.json"
         with open(config_path, "w") as f:
             json.dump(config_to_save, f, indent=2)
 
         return str(model_dir)
 
-    def test_toy_model_creation(self, qwen3_moe_toy_model_path):
+    def test_toy_model_creation(self, glm45_toy_model_path):
         """
         Test that the toy MoE model is created correctly and can be loaded.
 
         Args:
-            qwen3_moe_toy_model_path: Path to the toy Qwen3 MoE model (from fixture)
+            glm45_toy_model_path: Path to the toy GLM 4.5 MoE model (from fixture)
         """
         # Verify the model directory exists
-        model_path = Path(qwen3_moe_toy_model_path)
+        model_path = Path(glm45_toy_model_path)
         assert model_path.exists(), f"Model directory not found at {model_path}"
 
         # Check essential files exist
@@ -149,45 +157,44 @@ class TestQwen3MoEConversion:
         with open(config_file) as f:
             config_data = json.load(f)
 
-        assert config_data["model_type"] == "qwen3_moe"  # Qwen3 MoE uses qwen3_moe model type
-        assert config_data["hidden_size"] == 2048
-        assert config_data["num_hidden_layers"] == 2  # Updated to match toy config
-        assert config_data["num_attention_heads"] == 32
-        assert config_data["vocab_size"] == 151936
+        assert config_data["model_type"] == "glm"
+        assert config_data["hidden_size"] == 1024
+        assert config_data["intermediate_size"] == 2048
+        assert config_data["num_hidden_layers"] == 2
+        assert config_data["num_attention_heads"] == 8
+        assert config_data["vocab_size"] == 151552
         # Verify MoE specific parameters
-        assert config_data["num_experts"] == 4
+        assert config_data["n_routed_experts"] == 8
         assert config_data["num_experts_per_tok"] == 4
-        assert config_data["moe_intermediate_size"] == 768
+        assert config_data["moe_intermediate_size"] == 512
 
         # Try loading the model to verify it's valid
         try:
-            model = Qwen3MoeForCausalLM.from_pretrained(
-                qwen3_moe_toy_model_path,
+            from transformers import Glm4MoeForCausalLM
+
+            model = Glm4MoeForCausalLM.from_pretrained(
+                glm45_toy_model_path,
                 torch_dtype=torch.bfloat16,
                 low_cpu_mem_usage=False,  # Ensure full loading
+                trust_remote_code=True,
             )
-
-            # Try loading the tokenizer as well
-            try:
-                tokenizer = AutoTokenizer.from_pretrained(qwen3_moe_toy_model_path)
-                print(f"Tokenizer loaded successfully with vocab_size: {tokenizer.vocab_size}")
-            except Exception as e:
-                print(f"Warning: Could not load tokenizer (this might be OK for conversion testing): {e}")
 
             # Verify model structure
             assert hasattr(model, "model")
             assert hasattr(model.model, "layers")
-            assert len(model.model.layers) == 2  # num_hidden_layers updated to match toy config
+            assert len(model.model.layers) == 2  # num_hidden_layers
 
             # Verify MoE structure
-            first_layer = model.model.layers[0]
-            assert hasattr(first_layer, "mlp")
-            assert hasattr(first_layer.mlp, "experts")
-            assert len(first_layer.mlp.experts) == 4  # num_experts
+            # First layer is dense, second layer should have MoE structure
+            second_layer = model.model.layers[1]
+            assert hasattr(second_layer, "mlp")
+            # GLM 4.5 MoE structure check (may vary based on implementation)
+            if hasattr(second_layer.mlp, "experts"):
+                assert len(second_layer.mlp.experts) == 8  # n_routed_experts
 
-            print(f"SUCCESS: Qwen3 MoE toy model created and validated at {qwen3_moe_toy_model_path}")
+            print(f"SUCCESS: GLM 4.5 MoE toy model created and validated at {glm45_toy_model_path}")
             print("Model weights are correctly in bfloat16 format")
-            print(f"MoE structure validated: {len(first_layer.mlp.experts)} experts per layer")
+            print(f"MoE structure validated: {config_data['n_routed_experts']} experts")
 
         except Exception as e:
             assert False, f"Failed to load created toy MoE model: {e}"
@@ -201,12 +208,12 @@ class TestQwen3MoEConversion:
             (1, 1, 2, "EP"),
         ],
     )
-    def test_qwen3_moe_conversion_parallelism(self, qwen3_moe_toy_model_path, tmp_path, tp, pp, ep, test_name):
+    def test_glm45_conversion_parallelism(self, glm45_toy_model_path, tmp_path, tp, pp, ep, test_name):
         """
-        Test Qwen3 MoE model conversion with different parallelism configurations.
+        Test GLM 4.5 MoE model conversion with different parallelism configurations.
 
         Args:
-            qwen3_moe_toy_model_path: Path to the toy Qwen3 MoE model (from fixture)
+            glm45_toy_model_path: Path to the toy GLM 4.5 MoE model (from fixture)
             tmp_path: Pytest temporary path fixture
             tp: Tensor parallelism size
             pp: Pipeline parallelism size
@@ -215,7 +222,7 @@ class TestQwen3MoEConversion:
         """
 
         # Create temporary output directory for conversion results
-        test_output_dir = tmp_path / f"qwen3_moe_{test_name}"
+        test_output_dir = tmp_path / f"glm45_moe_{test_name}"
         test_output_dir.mkdir(exist_ok=True)
 
         # Run hf_megatron_roundtrip_multi_gpu.py with specified parallelism configuration on our toy MoE model
@@ -233,7 +240,7 @@ class TestQwen3MoEConversion:
             "--parallel-mode",
             "examples/conversion/hf_megatron_roundtrip_multi_gpu.py",
             "--hf-model-id",
-            qwen3_moe_toy_model_path,  # Use our local toy MoE model instead of downloading
+            glm45_toy_model_path,  # Use our local toy MoE model instead of downloading
             "--output-dir",
             str(test_output_dir),
             "--tp",
@@ -246,18 +253,18 @@ class TestQwen3MoEConversion:
 
         try:
             result = subprocess.run(
-                cmd, capture_output=True, text=True, cwd=Path(__file__).parent.parent.parent.parent
+                cmd, capture_output=True, text=True, cwd=Path(__file__).parent.parent.parent.parent.parent
             )
 
             # Check that the conversion completed successfully
             if result.returncode != 0:
                 print(f"STDOUT: {result.stdout}")
                 print(f"STDERR: {result.stderr}")
-                assert False, f"Qwen3 MoE {test_name} conversion failed with return code {result.returncode}"
+                assert False, f"GLM 4.5 MoE {test_name} conversion failed with return code {result.returncode}"
 
             # Verify that the converted model was saved
             # The output directory should be named after the last part of the model path
-            model_name = Path(qwen3_moe_toy_model_path).name  # "qwen3_moe_toy"
+            model_name = Path(glm45_toy_model_path).name  # "glm45_toy"
             converted_model_dir = test_output_dir / model_name
             assert converted_model_dir.exists(), f"Converted model directory not found at {converted_model_dir}"
 
@@ -280,26 +287,26 @@ class TestQwen3MoEConversion:
 
             assert weights_found, f"Model weights file not found in converted model at {converted_model_dir}"
 
-            # Verify the config contains Qwen3 MoE-specific parameters
+            # Verify the config contains GLM 4.5 MoE-specific parameters
             with open(config_file) as f:
                 saved_config = json.load(f)
 
-            assert saved_config["model_type"] == "qwen3_moe", (
-                "Model type should be qwen3_moe (Qwen3 MoE uses Qwen3MoeForCausalLM)"
+            assert saved_config["model_type"] == "glm", (
+                "Model type should be glm (GLM 4.5 MoE uses Glm4MoeForCausalLM)"
             )
-            assert saved_config["hidden_size"] == 2048, "Hidden size should match toy config"
-            assert saved_config["num_attention_heads"] == 32, "Number of attention heads should match toy config"
+            assert saved_config["hidden_size"] == 1024, "Hidden size should match toy config"
+            assert saved_config["num_attention_heads"] == 8, "Number of attention heads should match toy config"
             # Verify MoE specific parameters are preserved
-            assert saved_config["num_experts"] == 4, "Number of experts should match toy config"
+            assert saved_config["n_routed_experts"] == 8, "Number of routed experts should match toy config"
             assert saved_config["num_experts_per_tok"] == 4, "Number of experts per token should match toy config"
-            assert saved_config["moe_intermediate_size"] == 768, "MoE intermediate size should match toy config"
+            assert saved_config["moe_intermediate_size"] == 512, "MoE intermediate size should match toy config"
 
-            print(f"SUCCESS: Qwen3 MoE {test_name} conversion test completed successfully")
+            print(f"SUCCESS: GLM 4.5 MoE {test_name} conversion test completed successfully")
             print(f"Converted model saved at: {converted_model_dir}")
             print(
-                f"MoE parameters preserved: {saved_config['num_experts']} experts, {saved_config['num_experts_per_tok']} per token"
+                f"MoE parameters preserved: {saved_config['n_routed_experts']} experts, {saved_config['num_experts_per_tok']} per token"
             )
 
         except Exception as e:
-            print(f"Error during Qwen3 MoE {test_name} conversion test: {e}")
+            print(f"Error during GLM 4.5 MoE {test_name} conversion test: {e}")
             raise
