+++ tests/functional_tests/models/gpt_oss/test_gpt_oss_provider.py
