+++ tests/functional_tests/models/mistral/test_mistral_conversion.py
@@ -18,46 +18,43 @@ from pathlib import Path
 
 import pytest
 import torch
+from transformers import AutoTokenizer, MistralConfig, MistralForCausalLM
 
 
-HF_OLMOE_TOY_MODEL_CONFIG = {
-    "architectures": ["OlmoeForCausalLM"],
-    "attention_bias": False,
+HF_MISTRAL_TOY_MODEL_CONFIG = {
+    "architectures": ["MistralForCausalLM"],
     "attention_dropout": 0.0,
     "bos_token_id": 1,
     "eos_token_id": 2,
     "hidden_act": "silu",
-    "hidden_size": 1024,  # Smaller than real model for faster testing
+    "hidden_size": 4096,
     "initializer_range": 0.02,
-    "intermediate_size": 512,  # Reduced for testing
-    "max_position_embeddings": 4096,
-    "model_type": "olmoe",
-    "num_attention_heads": 16,
-    "num_experts": 8,  # Reduced from 64 for testing
-    "num_experts_per_tok": 4,  # Reduced from 8 for testing
-    "num_hidden_layers": 2,  # Much smaller for testing
-    "num_key_value_heads": 16,
-    "pad_token_id": 0,
+    "intermediate_size": 14336,
+    "max_position_embeddings": 32768,
+    "model_type": "mistral",
+    "num_attention_heads": 32,
+    "num_hidden_layers": 2,
+    "num_key_value_heads": 8,
     "rms_norm_eps": 1e-05,
-    "rope_theta": 10000.0,
-    "router_aux_loss_coef": 0.01,
+    "rope_theta": 1000000.0,
+    "sliding_window": None,
     "tie_word_embeddings": False,
     "torch_dtype": "bfloat16",
-    "transformers_version": "4.40.0",
+    "transformers_version": "4.40.1",
     "use_cache": True,
-    "vocab_size": 50304,
+    "vocab_size": 32768,
 }
 
 
-class TestOlMoEConversion:
+class TestMistralConversion:
     """
-    Test OLMoE model conversion from local HuggingFace model with different parallelism configurations.
+    Test Mistral model conversion from local HuggingFace model with different parallelism configurations.
     """
 
     @pytest.fixture(scope="class")
-    def olmoe_toy_model_path(self, tmp_path_factory):
+    def mistral_toy_model_path(self, tmp_path_factory):
         """
-        Create and save a HuggingFace OLMoE toy model from config to a temporary directory.
+        Create and save a HuggingFace Mistral toy model from config to a temporary directory.
 
         Args:
             tmp_path_factory: Pytest temporary path factory for class-scoped fixtures
@@ -66,32 +63,17 @@ class TestOlMoEConversion:
             str: Path to the saved HuggingFace model directory
         """
         # Create a temporary directory for this test class
-        temp_dir = tmp_path_factory.mktemp("olmoe_toy_model")
-        model_dir = temp_dir / "olmoe_toy"
+        temp_dir = tmp_path_factory.mktemp("mistral_toy_model")
+        model_dir = temp_dir / "mistral_toy"
 
-        # Create OLMoE config from the toy model config
-        try:
-            from transformers import OlmoeConfig, OlmoeForCausalLM
-
-            config = OlmoeConfig(**HF_OLMOE_TOY_MODEL_CONFIG)
-        except ImportError:
-            # If OlmoeConfig is not available, use AutoConfig
-            from transformers import AutoConfig
-
-            config = AutoConfig.for_model(model_type="olmoe", **HF_OLMOE_TOY_MODEL_CONFIG)
-
+        # Create Mistral config from the toy model config
+        config = MistralConfig(**HF_MISTRAL_TOY_MODEL_CONFIG)
+        for k, v in HF_MISTRAL_TOY_MODEL_CONFIG.items():
+            setattr(config, k, v)
         config.torch_dtype = torch.bfloat16  # Explicitly set the torch_dtype in config
 
         # Create model with random weights and convert to bfloat16
-        try:
-            from transformers import OlmoeForCausalLM
-
-            model = OlmoeForCausalLM(config)
-        except ImportError:
-            from transformers import AutoModelForCausalLM
-
-            model = AutoModelForCausalLM.from_config(config)
-
+        model = MistralForCausalLM(config)
         model = model.bfloat16()  # Use .bfloat16() method instead of .to()
 
         # Debug: Check model dtype before saving
@@ -99,41 +81,47 @@ class TestOlMoEConversion:
             print(f"Before save - {name}: {param.dtype}")
             break  # Just check the first parameter
 
-        # Download and save tokenizer from a reference model
-        # Try to use OLMoE tokenizer if available, otherwise fall back to GPT2
+        # Download and save tokenizer from a reference Mistral model
+        # We use the smallest available Mistral model for tokenizer artifacts
         try:
-            from transformers import AutoTokenizer
-
-            tokenizer = AutoTokenizer.from_pretrained("allenai/OLMoE-1B-7B-0125")
-        except Exception:
-            # Fallback to a generic tokenizer
-            from transformers import AutoTokenizer
-
             tokenizer = AutoTokenizer.from_pretrained("gpt2")
-            tokenizer.pad_token = tokenizer.eos_token
+            tokenizer.save_pretrained(model_dir)
+        except Exception as e:
+            print(f"Warning: Could not download tokenizer, creating minimal tokenizer files: {e}")
+            # Create minimal tokenizer files if download fails
+            # This is a fallback for offline environments
+            tokenizer_config = {
+                "tokenizer_class": "LlamaTokenizer",
+                "vocab_size": 32768,
+                "bos_token": "<s>",
+                "eos_token": "</s>",
+                "pad_token": None,
+                "unk_token": "<unk>",
+            }
 
-        tokenizer.save_pretrained(model_dir)
+            with open(model_dir / "tokenizer_config.json", "w") as f:
+                json.dump(tokenizer_config, f, indent=2)
 
         # Save model and config to directory
         model.save_pretrained(model_dir, safe_serialization=True)
 
         # Also save config.json explicitly to ensure compatibility with correct torch_dtype
-        config_to_save = HF_OLMOE_TOY_MODEL_CONFIG.copy()
+        config_to_save = HF_MISTRAL_TOY_MODEL_CONFIG.copy()
         config_path = model_dir / "config.json"
         with open(config_path, "w") as f:
             json.dump(config_to_save, f, indent=2)
 
         return str(model_dir)
 
-    def test_toy_model_creation(self, olmoe_toy_model_path):
+    def test_toy_model_creation(self, mistral_toy_model_path):
         """
         Test that the toy model is created correctly and can be loaded.
 
         Args:
-            olmoe_toy_model_path: Path to the toy OLMoE model (from fixture)
+            mistral_toy_model_path: Path to the toy Mistral model (from fixture)
         """
         # Verify the model directory exists
-        model_path = Path(olmoe_toy_model_path)
+        model_path = Path(mistral_toy_model_path)
         assert model_path.exists(), f"Model directory not found at {model_path}"
 
         # Check essential files exist
@@ -154,40 +142,37 @@ class TestOlMoEConversion:
         with open(config_file) as f:
             config_data = json.load(f)
 
-        assert config_data["model_type"] == "olmoe"
-        assert config_data["hidden_size"] == 1024
-        assert config_data["intermediate_size"] == 512
+        assert config_data["model_type"] == "mistral"
+        assert config_data["hidden_size"] == 4096
         assert config_data["num_hidden_layers"] == 2
-        assert config_data["num_attention_heads"] == 16
-        assert config_data["num_key_value_heads"] == 16
-        assert config_data["vocab_size"] == 50304
-        # Check OLMoE-specific MoE parameters
-        assert config_data["num_experts"] == 8
-        assert config_data["num_experts_per_tok"] == 4
-        assert config_data["router_aux_loss_coef"] == 0.01
+        assert config_data["num_attention_heads"] == 32
+        assert config_data["vocab_size"] == 32768
 
         # Try loading the model to verify it's valid
+        try:
+            model = MistralForCausalLM.from_pretrained(
+                mistral_toy_model_path,
+                torch_dtype=torch.bfloat16,
+                low_cpu_mem_usage=False,  # Ensure full loading
+            )
 
-        from transformers import OlmoeForCausalLM
+            # Try loading the tokenizer as well
+            try:
+                tokenizer = AutoTokenizer.from_pretrained(mistral_toy_model_path)
+                print(f"Tokenizer loaded successfully with vocab_size: {tokenizer.vocab_size}")
+            except Exception as e:
+                print(f"Warning: Could not load tokenizer (this might be OK for conversion testing): {e}")
 
-        model = OlmoeForCausalLM.from_pretrained(
-            olmoe_toy_model_path,
-            torch_dtype=torch.bfloat16,
-            low_cpu_mem_usage=False,  # Ensure full loading
-        )
+            # Verify model structure
+            assert hasattr(model, "model")
+            assert hasattr(model.model, "layers")
+            assert len(model.model.layers) == 2  # num_hidden_layers
 
-        from transformers import AutoTokenizer
+            print(f"SUCCESS: Toy model created and validated at {mistral_toy_model_path}")
+            print("Model weights are correctly in bfloat16 format")
 
-        tokenizer = AutoTokenizer.from_pretrained(olmoe_toy_model_path)
-        print(f"Tokenizer loaded successfully with vocab_size: {tokenizer.vocab_size}")
-
-        # Verify model structure
-        assert hasattr(model, "model")
-        assert hasattr(model.model, "layers")
-        assert len(model.model.layers) == 2  # num_hidden_layers
-
-        print(f"SUCCESS: Toy model created and validated at {olmoe_toy_model_path}")
-        print("Model weights are correctly in bfloat16 format")
+        except Exception as e:
+            assert False, f"Failed to load created toy model: {e}"
 
     @pytest.mark.run_only_on("GPU")
     @pytest.mark.parametrize(
@@ -197,21 +182,23 @@ class TestOlMoEConversion:
             (1, 2, "PP"),
         ],
     )
-    def test_olmoe_conversion_parallelism(self, olmoe_toy_model_path, tmp_path, tp, pp, test_name):
+    def test_mistral_conversion_parallelism(self, mistral_toy_model_path, tmp_path, tp, pp, test_name):
         """
-        Test OLMoE model conversion with different parallelism configurations.
+        Test Mistral model conversion with different parallelism configurations.
 
         Args:
-            olmoe_toy_model_path: Path to the toy OLMoE model (from fixture)
+            mistral_toy_model_path: Path to the toy Mistral model (from fixture)
             tmp_path: Pytest temporary path fixture
             tp: Tensor parallelism size
             pp: Pipeline parallelism size
             test_name: Name of the test for identification
         """
+
         # Create temporary output directory for conversion results
-        test_output_dir = tmp_path / f"olmoe_{test_name}"
+        test_output_dir = tmp_path / f"mistral_{test_name}"
         test_output_dir.mkdir(exist_ok=True)
 
+        # Run hf_megatron_roundtrip_multi_gpu.py with specified parallelism configuration on our toy model
         cmd = [
             "python",
             "-m",
@@ -221,12 +208,12 @@ class TestOlMoEConversion:
             "-m",
             "coverage",
             "run",
-            "--data-file=/workspace/.coverage",
-            "--source=/workspace/",
+            "--data-file=/opt/Megatron-Bridge/.coverage",
+            "--source=/opt/Megatron-Bridge/",
             "--parallel-mode",
             "examples/conversion/hf_megatron_roundtrip_multi_gpu.py",
             "--hf-model-id",
-            olmoe_toy_model_path,
+            mistral_toy_model_path,  # Use our local toy model instead of downloading
             "--output-dir",
             str(test_output_dir),
             "--tp",
@@ -237,18 +224,18 @@ class TestOlMoEConversion:
 
         try:
             result = subprocess.run(
-                cmd, capture_output=True, text=True, cwd=Path(__file__).parent.parent.parent.parent
+                cmd, capture_output=True, text=True, cwd=Path(__file__).parent.parent.parent.parent.parent
             )
 
             # Check that the conversion completed successfully
             if result.returncode != 0:
                 print(f"STDOUT: {result.stdout}")
                 print(f"STDERR: {result.stderr}")
-                assert False, f"OLMoE {test_name} conversion failed with return code {result.returncode}"
+                assert False, f"Mistral {test_name} conversion failed with return code {result.returncode}"
 
             # Verify that the converted model was saved
             # The output directory should be named after the last part of the model path
-            model_name = Path(olmoe_toy_model_path).name  # "olmoe_toy"
+            model_name = Path(mistral_toy_model_path).name  # "mistral_toy"
             converted_model_dir = test_output_dir / model_name
             assert converted_model_dir.exists(), f"Converted model directory not found at {converted_model_dir}"
 
@@ -263,23 +250,17 @@ class TestOlMoEConversion:
                 f"Model weights file not found in converted model at {converted_model_dir}"
             )
 
-            # Verify the config contains OLMoE-specific parameters
+            # Verify the config contains Mistral-specific parameters
             with open(config_file) as f:
                 saved_config = json.load(f)
 
-            assert saved_config["model_type"] == "olmoe", "Model type should be olmoe"
-            assert saved_config["hidden_size"] == 1024, "Hidden size should match toy config"
-            assert saved_config["intermediate_size"] == 512, "Intermediate size should match toy config"
-            assert saved_config["num_attention_heads"] == 16, "Number of attention heads should match toy config"
-            assert saved_config["num_key_value_heads"] == 16, "Number of key-value heads should match toy config"
-            # Verify OLMoE-specific MoE parameters
-            assert saved_config["num_experts"] == 8, "Number of experts should match"
-            assert saved_config["num_experts_per_tok"] == 4, "Number of experts per token should match"
-            assert saved_config["router_aux_loss_coef"] == 0.01, "Router aux loss coefficient should match"
+            assert saved_config["model_type"] == "mistral", "Model type should be mistral"
+            assert saved_config["hidden_size"] == 4096, "Hidden size should match toy config"
+            assert saved_config["num_attention_heads"] == 32, "Number of attention heads should match toy config"
 
-            print(f"SUCCESS: OLMoE {test_name} conversion test completed successfully")
+            print(f"SUCCESS: Mistral {test_name} conversion test completed successfully")
             print(f"Converted model saved at: {converted_model_dir}")
 
         except Exception as e:
-            print(f"Error during OLMoE {test_name} conversion test: {e}")
+            print(f"Error during Mistral {test_name} conversion test: {e}")
             raise
