+++ tests/functional_tests/models/mistral/test_mistral_provider.py
