+++ tests/functional_tests/models/nemotron/test_nemotron_conversion.py
@@ -13,51 +13,58 @@
 # limitations under the License.
 
 import json
-import os
-import shutil
 import subprocess
-import sys
 from pathlib import Path
 
 import pytest
 import torch
-from transformers import AutoConfig, AutoModelForCausalLM, AutoTokenizer
+from transformers import AutoTokenizer, NemotronConfig, NemotronForCausalLM
+from transformers.activations import ACT2FN
 
-from megatron.bridge.models.conversion.utils import get_causal_lm_class_via_auto_map
+
+# Register relu2 (squared_relu) activation function for Nemotron models
+def squared_relu(x):
+    """Squared ReLU activation function."""
+    return torch.pow(torch.nn.functional.relu(x), 2)
 
 
-# Overrides for 8B size
-HF_NEMOTRONH_TOY_MODEL_OVERRIDES = {
-    "attention_head_dim": 48,
-    "chunk_size": 48,
-    "expand": 2,
-    "hidden_size": 768,
-    "hybrid_override_pattern": "M*M-",
-    "initializer_range": 0.02,
-    "intermediate_size": 3072,
-    "layer_norm_epsilon": 1e-05,
-    "mamba_head_dim": 64,
-    "mamba_hidden_act": "silu",
-    "mamba_num_heads": 24,
-    "max_position_embeddings": 8192,
-    "n_groups": 8,
-    "num_attention_heads": 16,
-    "num_hidden_layers": 4,
-    "num_key_value_heads": 8,
-    "ssm_state_size": 128,
-    "vocab_size": 131072,
+# Register the activation function if not already present
+if "relu2" not in ACT2FN:
+    ACT2FN["relu2"] = squared_relu
+
+
+HF_NEMOTRON_TOY_MODEL_CONFIG = {
+    "architectures": ["NemotronForCausalLM"],
+    "bos_token_id": 1,
+    "eos_token_id": 2,
+    "hidden_act": "relu2",  # Nemotron default activation (squared_relu)
+    "hidden_size": 768,  # Smaller for toy model
+    "initializer_range": 0.0134,
+    "intermediate_size": 2304,  # 3 * hidden_size
+    "max_position_embeddings": 2048,  # Smaller for toy model
+    "model_type": "nemotron",
+    "norm_eps": 1e-05,
+    "num_attention_heads": 12,  # Smaller for toy model
+    "num_hidden_layers": 2,  # Very small for testing
+    "num_key_value_heads": 4,  # GQA with 4 KV heads
+    "partial_rotary_factor": 0.5,
+    "rope_theta": 10000.0,
+    "tie_word_embeddings": False,
+    "torch_dtype": "bfloat16",
+    "use_cache": True,
+    "vocab_size": 32000,  # Smaller vocab for toy model
 }
 
 
-class TestNemotronHConversion:
+class TestNemotronConversion:
     """
-    Test NemotronH model conversion from local HuggingFace model with different parallelism configurations.
+    Test Nemotron model conversion from local HuggingFace model with different parallelism configurations.
     """
 
     @pytest.fixture(scope="class")
-    def nemotronh_toy_model_path(self, tmp_path_factory):
+    def nemotron_toy_model_path(self, tmp_path_factory):
         """
-        Create and save a HuggingFace NemotronH toy model from config to a temporary directory.
+        Create and save a HuggingFace Nemotron toy model from config to a temporary directory.
 
         Args:
             tmp_path_factory: Pytest temporary path factory for class-scoped fixtures
@@ -66,45 +73,64 @@ class TestNemotronHConversion:
             str: Path to the saved HuggingFace model directory
         """
         # Create a temporary directory for this test class
-        temp_dir = tmp_path_factory.mktemp("nemotronh_toy_model")
-        model_dir = temp_dir / "nemotronh_toy"
+        temp_dir = tmp_path_factory.mktemp("nemotron_toy_model")
+        model_dir = temp_dir / "nemotron_toy"
 
-        # Create NemotronH toy model config by starting with 8B and applying overrides
-        # This avoids attempting import of NemotronHConfig from Transformers
-        config = AutoConfig.from_pretrained("nvidia/Nemotron-H-8B-Base-8K", trust_remote_code=True)
-        for k, v in HF_NEMOTRONH_TOY_MODEL_OVERRIDES.items():
-            setattr(config, k, v)
+        # Create Nemotron config from the toy model config
+        config = NemotronConfig(**HF_NEMOTRON_TOY_MODEL_CONFIG)
+        config.torch_dtype = torch.bfloat16  # Explicitly set the torch_dtype in config
 
         # Create model with random weights and convert to bfloat16
-        model_class = get_causal_lm_class_via_auto_map("nvidia/Nemotron-H-8B-Base-8K", config)
-        model = model_class(config)
-        model = model.bfloat16() if hasattr(model, "bfloat16") else model
+        model = NemotronForCausalLM(config)
+        model = model.bfloat16()  # Use .bfloat16() method instead of .to()
 
-        # Download and save tokenizer from a reference NemotronH model
-        tokenizer = AutoTokenizer.from_pretrained("nvidia/Nemotron-H-8B-Base-8K", trust_remote_code=True)
-        tokenizer.save_pretrained(model_dir)
+        # Debug: Check model dtype before saving
+        for name, param in model.named_parameters():
+            print(f"Before save - {name}: {param.dtype}")
+            break  # Just check the first parameter
 
-        # Save model, config, and modeling code to directory
+        # Create minimal tokenizer files
+        # Since Nemotron may not have a readily available tokenizer, we'll create minimal files
+        try:
+            # Try to use a compatible tokenizer (GPT-2 or similar)
+            tokenizer = AutoTokenizer.from_pretrained("gpt2")
+            tokenizer.save_pretrained(model_dir)
+        except Exception as e:
+            print(f"Warning: Could not download tokenizer, creating minimal tokenizer files: {e}")
+            # Create minimal tokenizer files if download fails
+            # This is a fallback for offline environments
+            tokenizer_config = {
+                "tokenizer_class": "GPT2Tokenizer",
+                "vocab_size": 32000,
+                "bos_token": "<s>",
+                "eos_token": "</s>",
+                "pad_token": "</s>",
+                "unk_token": "<unk>",
+            }
+
+            with open(model_dir / "tokenizer_config.json", "w") as f:
+                json.dump(tokenizer_config, f, indent=2)
+
+        # Save model and config to directory
         model.save_pretrained(model_dir, safe_serialization=True)
-        modeling_filepath = os.path.abspath(sys.modules[model_class.__module__].__file__)
-        shutil.copy(modeling_filepath, model_dir)
 
-        # Ensure config.json exists with expected keys
+        # Also save config.json explicitly to ensure compatibility with correct torch_dtype
+        config_to_save = HF_NEMOTRON_TOY_MODEL_CONFIG.copy()
         config_path = model_dir / "config.json"
         with open(config_path, "w") as f:
-            json.dump(model.config.to_dict(), f, indent=2)
+            json.dump(config_to_save, f, indent=2)
 
         return str(model_dir)
 
-    def test_toy_model_creation(self, nemotronh_toy_model_path):
+    def test_toy_model_creation(self, nemotron_toy_model_path):
         """
         Test that the toy model is created correctly and can be loaded.
 
         Args:
-            nemotronh_toy_model_path: Path to the toy NemotronH model (from fixture)
+            nemotron_toy_model_path: Path to the toy Nemotron model (from fixture)
         """
         # Verify the model directory exists
-        model_path = Path(nemotronh_toy_model_path)
+        model_path = Path(nemotron_toy_model_path)
         assert model_path.exists(), f"Model directory not found at {model_path}"
 
         # Check essential files exist
@@ -121,45 +147,38 @@ class TestNemotronHConversion:
         tokenizer_config_file = model_path / "tokenizer_config.json"
         assert tokenizer_config_file.exists(), f"tokenizer_config.json not found at {tokenizer_config_file}"
 
-        # Check for modeling file
-        nemotronh_modeling_file = model_path / "modeling_nemotron_h.py"
-        assert nemotronh_modeling_file.exists(), (
-            f"modeling_nemotron_h.py must be copied to toy model path. not found at {nemotronh_modeling_file}"
-        )
-
         # Load and verify config
         with open(config_file) as f:
             config_data = json.load(f)
 
-        assert config_data["model_type"] == "nemotron_h"
+        assert config_data["model_type"] == "nemotron"
         assert config_data["hidden_size"] == 768
-        assert config_data["intermediate_size"] == 3072
-        assert config_data["num_hidden_layers"] == 4  # Updated to match toy config
-        assert config_data["num_attention_heads"] == 16
-        assert config_data["vocab_size"] == 131072
+        assert config_data["num_hidden_layers"] == 2
+        assert config_data["num_attention_heads"] == 12
+        assert config_data["vocab_size"] == 32000
 
         # Try loading the model to verify it's valid
         try:
-            model = AutoModelForCausalLM.from_pretrained(
-                nemotronh_toy_model_path,
+            model = NemotronForCausalLM.from_pretrained(
+                nemotron_toy_model_path,
                 torch_dtype=torch.bfloat16,
                 low_cpu_mem_usage=False,  # Ensure full loading
-                trust_remote_code=True,
             )
 
             # Try loading the tokenizer as well
             try:
-                tokenizer = AutoTokenizer.from_pretrained(nemotronh_toy_model_path, trust_remote_code=True)
+                tokenizer = AutoTokenizer.from_pretrained(nemotron_toy_model_path)
                 print(f"Tokenizer loaded successfully with vocab_size: {tokenizer.vocab_size}")
             except Exception as e:
                 print(f"Warning: Could not load tokenizer (this might be OK for conversion testing): {e}")
 
             # Verify model structure
-            assert hasattr(model, "backbone")
-            assert hasattr(model.backbone, "layers")
-            assert len(model.backbone.layers) == 4  # num_hidden_layers updated to match toy config
+            assert hasattr(model, "model")
+            assert hasattr(model.model, "layers")
+            assert len(model.model.layers) == 2  # num_hidden_layers
 
-            print(f"SUCCESS: Toy model created and validated at {nemotronh_toy_model_path}")
+            print(f"SUCCESS: Toy model created and validated at {nemotron_toy_model_path}")
+            print("Model weights are correctly in bfloat16 format")
 
         except Exception as e:
             assert False, f"Failed to load created toy model: {e}"
@@ -172,12 +191,12 @@ class TestNemotronHConversion:
             (1, 2, "PP"),
         ],
     )
-    def test_nemotronh_conversion_parallelism(self, nemotronh_toy_model_path, tmp_path, tp, pp, test_name):
+    def test_nemotron_conversion_parallelism(self, nemotron_toy_model_path, tmp_path, tp, pp, test_name):
         """
-        Test NemotronH model conversion with different parallelism configurations.
+        Test Nemotron model conversion with different parallelism configurations.
 
         Args:
-            nemotronh_toy_model_path: Path to the toy NemotronH model (from fixture)
+            nemotron_toy_model_path: Path to the toy Nemotron model (from fixture)
             tmp_path: Pytest temporary path fixture
             tp: Tensor parallelism size
             pp: Pipeline parallelism size
@@ -185,10 +204,10 @@ class TestNemotronHConversion:
         """
 
         # Create temporary output directory for conversion results
-        test_output_dir = tmp_path / f"nemotronh_{test_name}"
+        test_output_dir = tmp_path / f"nemotron_{test_name}"
         test_output_dir.mkdir(exist_ok=True)
 
-        # Run hf_megatron_roundtrip_multi_gpu.py with specified parallelism configuration on our toy model
+        # Run multi_gpu_hf.py with specified parallelism configuration on our toy model
         cmd = [
             "python",
             "-m",
@@ -203,7 +222,7 @@ class TestNemotronHConversion:
             "--parallel-mode",
             "examples/conversion/hf_megatron_roundtrip_multi_gpu.py",
             "--hf-model-id",
-            nemotronh_toy_model_path,  # Use our local toy model instead of downloading
+            nemotron_toy_model_path,  # Use our local toy model instead of downloading
             "--output-dir",
             str(test_output_dir),
             "--tp",
@@ -214,18 +233,18 @@ class TestNemotronHConversion:
 
         try:
             result = subprocess.run(
-                cmd, capture_output=True, text=True, cwd=Path(__file__).parent.parent.parent.parent
+                cmd, capture_output=True, text=True, cwd=Path(__file__).parent.parent.parent.parent.parent
             )
 
             # Check that the conversion completed successfully
             if result.returncode != 0:
                 print(f"STDOUT: {result.stdout}")
                 print(f"STDERR: {result.stderr}")
-                assert False, f"NemotronH {test_name} conversion failed with return code {result.returncode}"
+                assert False, f"Nemotron {test_name} conversion failed with return code {result.returncode}"
 
             # Verify that the converted model was saved
             # The output directory should be named after the last part of the model path
-            model_name = Path(nemotronh_toy_model_path).name  # "nemotronh_toy"
+            model_name = Path(nemotron_toy_model_path).name  # "nemotron_toy"
             converted_model_dir = test_output_dir / model_name
             assert converted_model_dir.exists(), f"Converted model directory not found at {converted_model_dir}"
 
@@ -240,19 +259,17 @@ class TestNemotronHConversion:
                 f"Model weights file not found in converted model at {converted_model_dir}"
             )
 
-            # Verify the config contains NemotronH-specific parameters
+            # Verify the config contains Nemotron-specific parameters
             with open(config_file) as f:
                 saved_config = json.load(f)
 
-            assert saved_config["model_type"] == "nemotron_h", "Model type should be nemotron_h"
+            assert saved_config["model_type"] == "nemotron", "Model type should be nemotron"
             assert saved_config["hidden_size"] == 768, "Hidden size should match toy config"
-            assert saved_config["intermediate_size"] == 3072, "ffn hidden size should match toy config"
-            assert saved_config["num_hidden_layers"] == 4, "Number of hidden layers should match toy config"
-            assert saved_config["num_attention_heads"] == 16, "Number of attention heads should match toy config"
+            assert saved_config["num_attention_heads"] == 12, "Number of attention heads should match toy config"
 
-            print(f"SUCCESS: NemotronH {test_name} conversion test completed successfully")
+            print(f"SUCCESS: Nemotron {test_name} conversion test completed successfully")
             print(f"Converted model saved at: {converted_model_dir}")
 
         except Exception as e:
-            print(f"Error during NemotronH {test_name} conversion test: {e}")
+            print(f"Error during Nemotron {test_name} conversion test: {e}")
             raise
