+++ tests/functional_tests/models/nemotronh/test_nemotron_h_conversion.py
@@ -13,50 +13,51 @@
 # limitations under the License.
 
 import json
+import os
+import shutil
 import subprocess
+import sys
 from pathlib import Path
 
 import pytest
 import torch
-from transformers import GemmaConfig, GemmaForCausalLM, GemmaTokenizer
+from transformers import AutoConfig, AutoModelForCausalLM, AutoTokenizer
+
+from megatron.bridge.models.conversion.utils import get_causal_lm_class_via_auto_map
 
 
-HF_GEMMA_TOY_MODEL_CONFIG = {
-    "architectures": ["GemmaForCausalLM"],
-    "attention_bias": False,
-    "attention_dropout": 0.0,
-    "bos_token_id": 2,
-    "eos_token_id": 1,
-    "head_dim": 256,
-    "hidden_act": "gelu",
-    "hidden_size": 1024,  # Smaller than real 2B for faster testing
+# Overrides for 8B size
+HF_NEMOTRONH_TOY_MODEL_OVERRIDES = {
+    "attention_head_dim": 48,
+    "chunk_size": 48,
+    "expand": 2,
+    "hidden_size": 768,
+    "hybrid_override_pattern": "M*M-",
     "initializer_range": 0.02,
-    "intermediate_size": 4096,  # Smaller than real 2B for faster testing
+    "intermediate_size": 3072,
+    "layer_norm_epsilon": 1e-05,
+    "mamba_head_dim": 64,
+    "mamba_hidden_act": "silu",
+    "mamba_num_heads": 24,
     "max_position_embeddings": 8192,
-    "model_type": "gemma",
-    "num_attention_heads": 8,
-    "num_hidden_layers": 2,  # Much smaller for testing
-    "num_key_value_heads": 2,  # Changed from 1 to 2 to be divisible by TP=2
-    "pad_token_id": 0,
-    "rms_norm_eps": 1e-06,
-    "rope_scaling": None,
-    "rope_theta": 10000.0,
-    "torch_dtype": "bfloat16",
-    "transformers_version": "4.38.0.dev0",
-    "use_cache": True,
-    "vocab_size": 256000,
+    "n_groups": 8,
+    "num_attention_heads": 16,
+    "num_hidden_layers": 4,
+    "num_key_value_heads": 8,
+    "ssm_state_size": 128,
+    "vocab_size": 131072,
 }
 
 
-class TestGemmaConversion:
+class TestNemotronHConversion:
     """
-    Test Gemma model conversion from local HuggingFace model with different parallelism configurations.
+    Test NemotronH model conversion from local HuggingFace model with different parallelism configurations.
     """
 
     @pytest.fixture(scope="class")
-    def gemma_toy_model_path(self, tmp_path_factory):
+    def nemotronh_toy_model_path(self, tmp_path_factory):
         """
-        Create and save a HuggingFace Gemma toy model from config to a temporary directory.
+        Create and save a HuggingFace NemotronH toy model from config to a temporary directory.
 
         Args:
             tmp_path_factory: Pytest temporary path factory for class-scoped fixtures
@@ -65,56 +66,45 @@ class TestGemmaConversion:
             str: Path to the saved HuggingFace model directory
         """
         # Create a temporary directory for this test class
-        temp_dir = tmp_path_factory.mktemp("gemma_toy_model")
-        model_dir = temp_dir / "gemma_toy"
+        temp_dir = tmp_path_factory.mktemp("nemotronh_toy_model")
+        model_dir = temp_dir / "nemotronh_toy"
 
-        # Create Gemma config from the toy model config
-        config = GemmaConfig(**HF_GEMMA_TOY_MODEL_CONFIG)
-        config.torch_dtype = torch.bfloat16  # Explicitly set the torch_dtype in config
+        # Create NemotronH toy model config by starting with 8B and applying overrides
+        # This avoids attempting import of NemotronHConfig from Transformers
+        config = AutoConfig.from_pretrained("nvidia/Nemotron-H-8B-Base-8K", trust_remote_code=True)
+        for k, v in HF_NEMOTRONH_TOY_MODEL_OVERRIDES.items():
+            setattr(config, k, v)
 
         # Create model with random weights and convert to bfloat16
-        model = GemmaForCausalLM(config)
-        model = model.bfloat16()  # Use .bfloat16() method instead of .to()
+        model_class = get_causal_lm_class_via_auto_map("nvidia/Nemotron-H-8B-Base-8K", config)
+        model = model_class(config)
+        model = model.bfloat16() if hasattr(model, "bfloat16") else model
 
-        # Debug: Check model dtype before saving
-        for name, param in model.named_parameters():
-            print(f"Before save - {name}: {param.dtype}")
-            break  # Just check the first parameter
-
-        # Download and save tokenizer from a reference Gemma model
-        # We use the smallest available Gemma model for tokenizer artifacts
-        # First try to load from pre-mounted test data, then fall back to HuggingFace download
-        pre_downloaded_path = "/home/TestData/megatron_bridge/tokenizers/google/gemma-2b"
-        # Try loading from pre-downloaded location first
-        if Path(pre_downloaded_path).exists():
-            print(f"Loading tokenizer from pre-downloaded path: {pre_downloaded_path}")
-            tokenizer = GemmaTokenizer.from_pretrained(pre_downloaded_path)
-        else:
-            # Fall back to downloading from HuggingFace
-            print("Pre-downloaded tokenizer not found, attempting to download from HuggingFace")
-            tokenizer = GemmaTokenizer.from_pretrained("google/gemma-2b")
+        # Download and save tokenizer from a reference NemotronH model
+        tokenizer = AutoTokenizer.from_pretrained("nvidia/Nemotron-H-8B-Base-8K", trust_remote_code=True)
         tokenizer.save_pretrained(model_dir)
 
-        # Save model and config to directory
+        # Save model, config, and modeling code to directory
         model.save_pretrained(model_dir, safe_serialization=True)
+        modeling_filepath = os.path.abspath(sys.modules[model_class.__module__].__file__)
+        shutil.copy(modeling_filepath, model_dir)
 
-        # Also save config.json explicitly to ensure compatibility with correct torch_dtype
-        config_to_save = HF_GEMMA_TOY_MODEL_CONFIG.copy()
+        # Ensure config.json exists with expected keys
         config_path = model_dir / "config.json"
         with open(config_path, "w") as f:
-            json.dump(config_to_save, f, indent=2)
+            json.dump(model.config.to_dict(), f, indent=2)
 
         return str(model_dir)
 
-    def test_toy_model_creation(self, gemma_toy_model_path):
+    def test_toy_model_creation(self, nemotronh_toy_model_path):
         """
         Test that the toy model is created correctly and can be loaded.
 
         Args:
-            gemma_toy_model_path: Path to the toy Gemma model (from fixture)
+            nemotronh_toy_model_path: Path to the toy NemotronH model (from fixture)
         """
         # Verify the model directory exists
-        model_path = Path(gemma_toy_model_path)
+        model_path = Path(nemotronh_toy_model_path)
         assert model_path.exists(), f"Model directory not found at {model_path}"
 
         # Check essential files exist
@@ -131,40 +121,45 @@ class TestGemmaConversion:
         tokenizer_config_file = model_path / "tokenizer_config.json"
         assert tokenizer_config_file.exists(), f"tokenizer_config.json not found at {tokenizer_config_file}"
 
+        # Check for modeling file
+        nemotronh_modeling_file = model_path / "modeling_nemotron_h.py"
+        assert nemotronh_modeling_file.exists(), (
+            f"modeling_nemotron_h.py must be copied to toy model path. not found at {nemotronh_modeling_file}"
+        )
+
         # Load and verify config
         with open(config_file) as f:
             config_data = json.load(f)
 
-        assert config_data["model_type"] == "gemma"
-        assert config_data["hidden_size"] == 1024
-        assert config_data["num_hidden_layers"] == 2
-        assert config_data["num_attention_heads"] == 8
-        assert config_data["num_key_value_heads"] == 2
-        assert config_data["vocab_size"] == 256000
-        assert config_data["head_dim"] == 256
+        assert config_data["model_type"] == "nemotron_h"
+        assert config_data["hidden_size"] == 768
+        assert config_data["intermediate_size"] == 3072
+        assert config_data["num_hidden_layers"] == 4  # Updated to match toy config
+        assert config_data["num_attention_heads"] == 16
+        assert config_data["vocab_size"] == 131072
 
         # Try loading the model to verify it's valid
         try:
-            model = GemmaForCausalLM.from_pretrained(
-                gemma_toy_model_path,
+            model = AutoModelForCausalLM.from_pretrained(
+                nemotronh_toy_model_path,
                 torch_dtype=torch.bfloat16,
                 low_cpu_mem_usage=False,  # Ensure full loading
+                trust_remote_code=True,
             )
 
             # Try loading the tokenizer as well
             try:
-                tokenizer = GemmaTokenizer.from_pretrained(gemma_toy_model_path)
+                tokenizer = AutoTokenizer.from_pretrained(nemotronh_toy_model_path, trust_remote_code=True)
                 print(f"Tokenizer loaded successfully with vocab_size: {tokenizer.vocab_size}")
             except Exception as e:
                 print(f"Warning: Could not load tokenizer (this might be OK for conversion testing): {e}")
 
             # Verify model structure
-            assert hasattr(model, "model")
-            assert hasattr(model.model, "layers")
-            assert len(model.model.layers) == 2  # num_hidden_layers
+            assert hasattr(model, "backbone")
+            assert hasattr(model.backbone, "layers")
+            assert len(model.backbone.layers) == 4  # num_hidden_layers updated to match toy config
 
-            print(f"SUCCESS: Toy model created and validated at {gemma_toy_model_path}")
-            print("Model weights are correctly in bfloat16 format")
+            print(f"SUCCESS: Toy model created and validated at {nemotronh_toy_model_path}")
 
         except Exception as e:
             assert False, f"Failed to load created toy model: {e}"
@@ -177,12 +172,12 @@ class TestGemmaConversion:
             (1, 2, "PP"),
         ],
     )
-    def test_gemma_conversion_parallelism(self, gemma_toy_model_path, tmp_path, tp, pp, test_name):
+    def test_nemotronh_conversion_parallelism(self, nemotronh_toy_model_path, tmp_path, tp, pp, test_name):
         """
-        Test Gemma model conversion with different parallelism configurations.
+        Test NemotronH model conversion with different parallelism configurations.
 
         Args:
-            gemma_toy_model_path: Path to the toy Gemma model (from fixture)
+            nemotronh_toy_model_path: Path to the toy NemotronH model (from fixture)
             tmp_path: Pytest temporary path fixture
             tp: Tensor parallelism size
             pp: Pipeline parallelism size
@@ -190,9 +185,10 @@ class TestGemmaConversion:
         """
 
         # Create temporary output directory for conversion results
-        test_output_dir = tmp_path / f"gemma_{test_name}"
+        test_output_dir = tmp_path / f"nemotronh_{test_name}"
         test_output_dir.mkdir(exist_ok=True)
 
+        # Run hf_megatron_roundtrip_multi_gpu.py with specified parallelism configuration on our toy model
         cmd = [
             "python",
             "-m",
@@ -207,7 +203,7 @@ class TestGemmaConversion:
             "--parallel-mode",
             "examples/conversion/hf_megatron_roundtrip_multi_gpu.py",
             "--hf-model-id",
-            gemma_toy_model_path,
+            nemotronh_toy_model_path,  # Use our local toy model instead of downloading
             "--output-dir",
             str(test_output_dir),
             "--tp",
@@ -218,17 +214,18 @@ class TestGemmaConversion:
 
         try:
             result = subprocess.run(
-                cmd, capture_output=True, text=True, cwd=Path(__file__).parent.parent.parent.parent
+                cmd, capture_output=True, text=True, cwd=Path(__file__).parent.parent.parent.parent.parent
             )
+
             # Check that the conversion completed successfully
             if result.returncode != 0:
                 print(f"STDOUT: {result.stdout}")
                 print(f"STDERR: {result.stderr}")
-                assert False, f"Gemma {test_name} conversion failed with return code {result.returncode}"
+                assert False, f"NemotronH {test_name} conversion failed with return code {result.returncode}"
 
             # Verify that the converted model was saved
             # The output directory should be named after the last part of the model path
-            model_name = Path(gemma_toy_model_path).name  # "gemma_toy"
+            model_name = Path(nemotronh_toy_model_path).name  # "nemotronh_toy"
             converted_model_dir = test_output_dir / model_name
             assert converted_model_dir.exists(), f"Converted model directory not found at {converted_model_dir}"
 
@@ -243,19 +240,19 @@ class TestGemmaConversion:
                 f"Model weights file not found in converted model at {converted_model_dir}"
             )
 
-            # Verify the config contains Gemma-specific parameters
+            # Verify the config contains NemotronH-specific parameters
             with open(config_file) as f:
                 saved_config = json.load(f)
 
-            assert saved_config["model_type"] == "gemma", "Model type should be gemma"
-            assert saved_config["hidden_size"] == 1024, "Hidden size should match toy config"
-            assert saved_config["num_attention_heads"] == 8, "Number of attention heads should match toy config"
-            assert saved_config["num_key_value_heads"] == 2, "Number of key-value heads should match toy config"
-            assert saved_config["head_dim"] == 256, "Head dimension should match toy config"
+            assert saved_config["model_type"] == "nemotron_h", "Model type should be nemotron_h"
+            assert saved_config["hidden_size"] == 768, "Hidden size should match toy config"
+            assert saved_config["intermediate_size"] == 3072, "ffn hidden size should match toy config"
+            assert saved_config["num_hidden_layers"] == 4, "Number of hidden layers should match toy config"
+            assert saved_config["num_attention_heads"] == 16, "Number of attention heads should match toy config"
 
-            print(f"SUCCESS: Gemma {test_name} conversion test completed successfully")
+            print(f"SUCCESS: NemotronH {test_name} conversion test completed successfully")
             print(f"Converted model saved at: {converted_model_dir}")
 
         except Exception as e:
-            print(f"Error during Gemma {test_name} conversion test: {e}")
+            print(f"Error during NemotronH {test_name} conversion test: {e}")
             raise
