+++ tests/functional_tests/models/qwen/test_qwen2_conversion.py
@@ -224,7 +224,7 @@ class TestQwen2Conversion:
 
         try:
             result = subprocess.run(
-                cmd, capture_output=True, text=True, cwd=Path(__file__).parent.parent.parent.parent
+                cmd, capture_output=True, text=True, cwd=Path(__file__).parent.parent.parent.parent.parent
             )
 
             # Check that the conversion completed successfully
