+++ tests/functional_tests/models/qwen/test_qwen2_provider.py
