+++ tests/functional_tests/models/qwen/test_qwen3_moe_provider.py
