+++ tests/functional_tests/models/qwen/test_qwen3_next_provider.py
