+++ tests/functional_tests/models/qwen/test_qwen3_provider.py
